# Refactor API for unconventionally-indexed arrays