… base/arraymath.jl
… base/broadcast.jl
… base/deprecated.jl
… base/essentials.jl
… base/multidimensional.jl
… base/number.jl
… base/operators.jl
… base/permuteddimsarray.jl
… base/range.jl
… base/reshapedarray.jl
… base/serialize.jl
… base/sort.jl
… base/sparse/sparse.jl
… base/sparse/sparsematrix.jl
… base/subarray.jl
… doc/manual/interfaces.rst
… doc/stdlib/arrays.rst
… test/broadcast.jl
… test/choosetests.jl
… test/offsetarray.jl
… test/ranges.jl
… test/subarray.jl