This is effectively "cleanup" from #16260. It gets rid of `shape` and renames `allocate_for` to be `similar`. It achieves these by introducing a new Range type, `OneTo(n)`, which is equivalent (in a value-sense) to `1:n`. These are changes recommended by @JeffBezanson. I'm hoping to avoid having `OneTo` "leak" out into user space, so it's not exported, but I suspect we may need to export it at some point (let's first see how far we can get without doing that).

Despite the name of this branch, it does not make unconventional indices safer (ref `@arraysafe` in #16973); it seems better to do that in a separate PR.

A key thing here is going to be to find out whether I've done something nasty for performance, so @nanosoldier `runbenchmarks(ALL, vs=:master)`.