@timholy 2016-06-27
I seem to have messed up the nanosoldier call, so let's try again after adding quotes: @nanosoldier `runbenchmarks(ALL, vs=":master")`.