@nanosoldier 2016-06-27
[Your benchmark job](https://github.com/JuliaLang/julia/pull/17137#issuecomment-228703462) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/75370acde884780e55a4b669a4ac2dc0e6bd45a8/b1b77a1_vs_a543e0c/report.md). cc @jrevels