@timholy 2016-06-29
I've canceled the CI here; while the runtime looks good, 1b684d5 ("eliminate the `indicesbehavior` trait") absolutely hammers compile time.

Master:
```jl
julia> using Base.Test

shell> cd test/linalg/
/home/tim/src/julia-0.5b/test/linalg

julia> @time include("schur.jl")
 17.479364 seconds (13.13 M allocations: 516.591 MB, 1.47% gc time)

julia> @time include("schur.jl")
  1.011176 seconds (445.60 k allocations: 23.307 MB, 1.33% gc time)
```

This PR:
```jl
julia> using Base.Test

shell> cd test/linalg/
/home/tim/src/julia-0.5/test/linalg

julia> @time include("schur.jl")
 44.836752 seconds (72.31 M allocations: 3.040 GB, 2.34% gc time)

julia> @time include("schur.jl")
  0.902519 seconds (447.55 k allocations: 23.480 MB, 1.02% gc time)
```
I'll see what I can figure out. Meanwhile, it's a nice test case for folks who like to look at compiler performance :wink: .