@timholy 2016-06-29
Wow, the ProfileView looks like cathedrals of colorful `inference` sand, reaching towards the heavens. So pretty and spiritual, I think I'll frame it.
![image](https://cloud.githubusercontent.com/assets/1525481/16467802/97208b60-3e0e-11e6-952c-af3b4e094207.png)