@timholy 2016-06-29
Interesting observation: if I comment out [these](https://github.com/JuliaLang/julia/blob/70a7da94a4e3a28819e61992b929f09f5a1f363d/test/linalg/schur.jl#L19) [loops](https://github.com/JuliaLang/julia/blob/70a7da94a4e3a28819e61992b929f09f5a1f363d/test/linalg/schur.jl#L23) and just manually set the values, then master and this PR have similar compile times. But set `eltya = Int` and leave the `atype` loop intact, and there's a nearly 3x difference in compile times. Also possibly-relevant, changing the `for` loop to a `while` loop,
```diff
diff --git a/test/linalg/schur.jl b/test/linalg/schur.jl
index 5370afd..02a07dc 100644
--- a/test/linalg/schur.jl
+++ b/test/linalg/schur.jl
@@ -16,11 +16,17 @@ srand(1234321)
 areal = randn(n,n)/2
 aimg  = randn(n,n)/2
 
-for eltya in (Float32, Float64, Complex64, Complex128, Int)
+# for eltya in (Float32, Float64, Complex64, Complex128, Int)
+eltya = Int
     a = eltya == Int ? rand(1:7, n, n) : convert(Matrix{eltya}, eltya <: Complex ? complex(areal, aimg) : areal)
     asym = a'+a                  # symmetric indefinite
     apd  = a'*a                 # symmetric positive-definite
-    for atype in ("Array", "SubArray")
+    indx = 1
+    atypes = ["Array", "SubArray"]
+    while indx <= length(atypes)
+#    for atype in ("Array", "SubArray")
+        atype = atypes[indx]
+        indx += 1
         if atype == "Array"
             a = a
         else
@@ -96,4 +102,4 @@ for eltya in (Float32, Float64, Complex64, Complex128, Int)
         @test NS[:S] ≈ sS
         @test NS[:Z] ≈ sZ
     end
-end
+# end
```
does **not** fix the performance, so this seems different from #16122.