@timholy 2016-06-29
Here's a minimal reproducer:
```jl
using Base.Test

a = rand(1:7, 10, 10)
for atype in ("Array", "SubArray")
    d,v = eig(a)
end
```
Master:
```jl
julia> @time include("/tmp/tim/schur.jl")
  1.740690 seconds (1.93 M allocations: 84.396 MB, 2.91% gc time)
```
This PR:
```jl
julia> @time include("/tmp/tim/schur.jl")
 13.267818 seconds (46.09 M allocations: 1.985 GB, 5.30% gc time)
```
But now comment out the `for` loop, and they are the same.