@JeffBezanson 2016-06-29
Wow, interesting. What happens if you use an array instead of a tuple: `for atype in ["Array", "SubArray"]`?