@timholy 2016-06-29
> Wow, interesting.

In the Chinese curse-sense, yes.  :smile:

> What happens if you use an array instead of a tuple: for atype in ["Array", "SubArray"]?

Still slow.

I've debugged this a little further; one relevant point is that I can call 
```
a = rand(1:7, 10, 10)
eig(a)
```
before `include`ing the file with the loop (or copy/pasting into the REPL), and it's *still* dirt-slow. But I can't take the `eig` call out of the loop without making it fast. Finally, it's all one giant toplevel-inference call: this diff
```diff
+
+type InfRef
+    val::Bool
+end
+const debug = InfRef(false)
 function typeinf_ext(linfo::LambdaInfo)
     if isdefined(linfo, :def)
+        local tstart
+        if debug.val
+            print(linfo.def.name, " start: ", ccall(:jl_clock_now, Float64, ()))
+        end
         # method lambda - infer this specialization via the method cache
         (code, _t, inferred) = typeinf_edge(linfo.def, linfo.specTypes, linfo.sparam_vals, true, true, true, linfo)
         if inferred && code.inferred && linfo !== code
@@ -1567,13 +1576,22 @@ function typeinf_ext(linfo::LambdaInfo)
             linfo.inferred = true
             linfo.inInference = false
         end
+        if debug.val
+            print(linfo.def.name, " stop: ", ccall(:jl_clock_now, Float64, ()))
+        end
         return code
     else
         # toplevel lambda - infer directly
+        if debug.val
+            println("toplevel inference")
+        end
         linfo.inInference = true
         frame = InferenceState(linfo, true, true)
         typeinf_loop(frame)
         @assert frame.inferred # TODO: deal with this better
+        if debug.val
+            println("toplevel inference done")
+        end
         return linfo
     end
 end
```
gives this output (having already called `eig` once):
```
julia> for atype in ("Array", "SubArray")
           d,v = eig(a)
       end
toplevel inference
toplevel inference done
:put! start: 1.46724e+09:put! stop: 1.46724e+09:start start: 1.46724e+09:start stop: 1.46724e+09:indexed_next start: 1.46724e+09:indexed_next stop: 1.46724e+09
```
and all the waiting is before the "toplevel inference done".